requirements.txt
Django>=2.2,<2.3
wagtail>=2.7,<2.8
django-rest-framework>=0.1.0
